Fixed circular import - flame length still missing
